requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==2.4.2
Django==2.1.4
django-extensions==2.1.4
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.4.0
pycparser==2.19
pylint==2.2.2
pyOpenSSL==18.0.0
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.1.1
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
